requirements.txt
fastapi
uvicorn[standard]
groq
pydantic
pydantic-settings
python-dotenv
requests
sentence-transformers
faiss-cpu
numpy
